requirements.txt
langchain-community
aiohttp
python-dotenv
langchain_groq
nest_asyncio
html2text
bs4